Separate user response handling into its own service
